El player rota con el mouse

+++ Server/src/ClientEventHandlers/Game/RotateHandler.cpp
@@ -3,7 +3,7 @@
 RotateHandler::RotateHandler() {}
 
 void RotateHandler::handle(Event& event, GameWorld& world){
-	int angle = *((int*)&(event.get_msg()[1]));
+	int angle = (int)*((uint16_t*)&(event.get_msg()[1]));
 	world.get_character(event.get_id())
 		 .set_angle(angle);
 }
